50-60% of the time spent in det() function, when constructing new facets, but qconvex not

+++ src/quickhull.cpp
@@ -64,7 +64,7 @@ main(int argc, char * argv[])
         err_ << "io: count line" << std::endl;
         return EXIT_FAILURE;
     }
-    using value_type = double;
+    using value_type = float;
     using point = std::valarray< value_type >;
     using points = std::vector< point >;
     size_type count_ = 0;
